# api/api.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description:  |-
    Спецификация API для VK TECH
  version: 1.0.0
...
paths:
...
  /user/ChangeUserData:
    put:
      tags:
        - user
      summary: Проверка куки и изменение данных о себе
      operationId: ChangeUserData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserData'
      responses:
        '200':
          description: Successful operation
          headers:
            Set-Cookie:
              description: 0JfQsNC60L7QtNC40YDQvtCy0LDQvdC90YvQtSDQutGD0LrQuA==
              schema:
                type: string
        '405':
          description: Method Not Allowed
        '500':
          description: Server Error
...
  /films/GetAllFilms:
    get:
      tags:
        - film
      summary: Получние списка фильмов с актерами
      operationId: GetAllFilms
      parameters:
      - name: session_token
        in: header
        description: Куки пользователя
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetAllFilms'

        '405':
          description: Method Not Allowed
        '500':
          description: Server Error
  /films/ChangeFilmInfo:
      put:
        tags:
          - film
        summary: Изменение  информации о фильме
        operationId: ChangeFilmInfo
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeFilmInfo'
        parameters:
        - name: session_token
          in: header
          description: Куки пользователя
          required: true
          schema:
            type: string
        responses:
          '200':
            description: Successful operation
          '400':
            description: Wrong Data
          '405':
            description: Method Not Allowed
          '500':
            description: Server Error
...
  /films/DeleteFilm:
      delete:
        tags:
          - film
        summary: Изменение  информации о фильме
        operationId: DeleteFilm
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteFilm'
        parameters:
        - name: session_token
          in: header
          description: Куки пользователя
          required: true
          schema:
            type: string
        responses:
          '200':
            description: Successful operation
          '400':
            description: Wrong Data
          '405':
            description: Method Not Allowed
          '500':
            description: Server Error
...
components:
  securitySchemes:
    cookieAuth:
      name: session_token
      type: apiKey
      in: cookie
  schemas:
...
    ChangeUserData:
      type:  object
      properties:
        login:
          type: string
          description: Имя пользователя
          example: Cringe
        password:
          type: string
          description: Пароль пользователя
          example: 1234qwerty
      required:
        - login
        - password
...
    GetAllFilms:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор актера
          example:  1
        name:
          type: string
          description: Название фильма
          example:  Drive
        description:
          type: string
          description: Описание фильма
          example: A brilliant driver - by day he performs stunt tricks on Hollywood movie sets, and by night he leads a risky game. But one dangerous contract - and a reward is put on his life. Now, to stay alive and save his charming neighbor, he must do what he does best - skillfully evade pursuit.
        enterdate:
          type: string
          format: date-time
          description: Дата рождения актера
          example:  2011-05-20
        rate:
          type:   integer
          description: Рейтинг фильма
          example:  7.5
        score:
          type: integer
          description: Сколько всего баллов набрал фильм
          example: 75
        votes: 
          type: integer
          description: Сколько всего  голосовало
          example: 10
        acters:
          type: array
          items:
            type: integer
          description: Список идентификаторов актеров, которые снялись в фильме
          example: [1, 2, 3]
    ChangeFilmInfo:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор актера
          example:  1
        name:
          type: string
          description: Название фильма
          example:  Drive
        description:
          type: string
          description: Описание фильма
          example: A brilliant driver - by day he performs stunt tricks on Hollywood movie sets, and by night he leads a risky game. But one dangerous contract - and a reward is put on his life. Now, to stay alive and save his charming neighbor, he must do what he does best - skillfully evade pursuit.
        enterdate:
          type: string
          format: date-time
          description: Дата рождения актера
          example:  2011-05-20
        score:
          type: integer
          description: Сколько всего баллов набрал фильм
          example: 75
        votes: 
          type: integer
          description: Сколько всего  голосовало
          example: 10
        acters:
          type: array
          items:
            type: integer
          description: Список идентификаторов актеров, которые снялись в фильме
          example: [1, 2, 3]
      required:
        - id
        - name
        - description
        - enterdate
        - score
        - votes
        - acters
...
    DeleteFilm:
      type: object
      properties:
        id:
          type: integer
          description:  ID фильма
          example: 1
      required:
        - id